@@ -26,9 +26,9 @@ public class DefaultCatalog extends CatalogImpl {
     public static final DefaultCatalog DEFAULT_CATALOG = new DefaultCatalog();
 
     /**
-     * The schema <code>minance</code>.
+     * The schema <code>DEFAULT_SCHEMA</code>.
      */
-    public final Minance MINANCE = Minance.MINANCE;
+    public final DefaultSchema DEFAULT_SCHEMA = DefaultSchema.DEFAULT_SCHEMA;
 
     /**
      * No further instances allowed
@@ -40,7 +40,7 @@ public class DefaultCatalog extends CatalogImpl {
     @Override
     public final List<Schema> getSchemas() {
         return Arrays.asList(
-            Minance.MINANCE
+            DefaultSchema.DEFAULT_SCHEMA
         );
     }
 
