@@ -34,28 +34,28 @@ public class MinanceCategory implements Serializable {
     }
 
     /**
-     * Getter for <code>minance.minance_category.m_category_id</code>.
+     * Getter for <code>minance_category.m_category_id</code>.
      */
     public Integer getMCategoryId() {
         return this.mCategoryId;
     }
 
     /**
-     * Setter for <code>minance.minance_category.m_category_id</code>.
+     * Setter for <code>minance_category.m_category_id</code>.
      */
     public void setMCategoryId(Integer mCategoryId) {
         this.mCategoryId = mCategoryId;
     }
 
     /**
-     * Getter for <code>minance.minance_category.category</code>.
+     * Getter for <code>minance_category.category</code>.
      */
     public String getCategory() {
         return this.category;
     }
 
     /**
-     * Setter for <code>minance.minance_category.category</code>.
+     * Setter for <code>minance_category.category</code>.
      */
     public void setCategory(String category) {
         this.category = category;
