@@ -20,260 +20,260 @@ import today.ihelio.jooq.tables.Transactions;
  * This class is generated by jOOQ.
  */
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class TransactionsRecord extends UpdatableRecordImpl<TransactionsRecord> implements Record18<Integer, Integer, String, String, String, LocalDate, LocalDate, String, String, String, String, String, String, BigDecimal, String, String, String, Boolean> {
+public class TransactionsRecord extends UpdatableRecordImpl<TransactionsRecord> implements Record18<Integer, Integer, String, String, String, LocalDate, LocalDate, String, String, String, String, String, String, BigDecimal, String, String, String, Integer> {
 
     private static final long serialVersionUID = 1L;
 
     /**
-     * Setter for <code>minance.transactions.transaction_id</code>.
+     * Setter for <code>transactions.transaction_id</code>.
      */
     public void setTransactionId(Integer value) {
         set(0, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.transaction_id</code>.
+     * Getter for <code>transactions.transaction_id</code>.
      */
     public Integer getTransactionId() {
         return (Integer) get(0);
     }
 
     /**
-     * Setter for <code>minance.transactions.account_id</code>.
+     * Setter for <code>transactions.account_id</code>.
      */
     public void setAccountId(Integer value) {
         set(1, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.account_id</code>.
+     * Getter for <code>transactions.account_id</code>.
      */
     public Integer getAccountId() {
         return (Integer) get(1);
     }
 
     /**
-     * Setter for <code>minance.transactions.category</code>.
+     * Setter for <code>transactions.category</code>.
      */
     public void setCategory(String value) {
         set(2, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.category</code>.
+     * Getter for <code>transactions.category</code>.
      */
     public String getCategory() {
         return (String) get(2);
     }
 
     /**
-     * Setter for <code>minance.transactions.description</code>.
+     * Setter for <code>transactions.description</code>.
      */
     public void setDescription(String value) {
         set(3, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.description</code>.
+     * Getter for <code>transactions.description</code>.
      */
     public String getDescription() {
         return (String) get(3);
     }
 
     /**
-     * Setter for <code>minance.transactions.transaction_type</code>.
+     * Setter for <code>transactions.transaction_type</code>.
      */
     public void setTransactionType(String value) {
         set(4, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.transaction_type</code>.
+     * Getter for <code>transactions.transaction_type</code>.
      */
     public String getTransactionType() {
         return (String) get(4);
     }
 
     /**
-     * Setter for <code>minance.transactions.transaction_date</code>.
+     * Setter for <code>transactions.transaction_date</code>.
      */
     public void setTransactionDate(LocalDate value) {
         set(5, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.transaction_date</code>.
+     * Getter for <code>transactions.transaction_date</code>.
      */
     public LocalDate getTransactionDate() {
         return (LocalDate) get(5);
     }
 
     /**
-     * Setter for <code>minance.transactions.post_date</code>.
+     * Setter for <code>transactions.post_date</code>.
      */
     public void setPostDate(LocalDate value) {
         set(6, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.post_date</code>.
+     * Getter for <code>transactions.post_date</code>.
      */
     public LocalDate getPostDate() {
         return (LocalDate) get(6);
     }
 
     /**
-     * Setter for <code>minance.transactions.memo</code>.
+     * Setter for <code>transactions.memo</code>.
      */
     public void setMemo(String value) {
         set(7, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.memo</code>.
+     * Getter for <code>transactions.memo</code>.
      */
     public String getMemo() {
         return (String) get(7);
     }
 
     /**
-     * Setter for <code>minance.transactions.address</code>.
+     * Setter for <code>transactions.address</code>.
      */
     public void setAddress(String value) {
         set(8, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.address</code>.
+     * Getter for <code>transactions.address</code>.
      */
     public String getAddress() {
         return (String) get(8);
     }
 
     /**
-     * Setter for <code>minance.transactions.city</code>.
+     * Setter for <code>transactions.city</code>.
      */
     public void setCity(String value) {
         set(9, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.city</code>.
+     * Getter for <code>transactions.city</code>.
      */
     public String getCity() {
         return (String) get(9);
     }
 
     /**
-     * Setter for <code>minance.transactions.state_name</code>.
+     * Setter for <code>transactions.state_name</code>.
      */
     public void setStateName(String value) {
         set(10, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.state_name</code>.
+     * Getter for <code>transactions.state_name</code>.
      */
     public String getStateName() {
         return (String) get(10);
     }
 
     /**
-     * Setter for <code>minance.transactions.country</code>.
+     * Setter for <code>transactions.country</code>.
      */
     public void setCountry(String value) {
         set(11, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.country</code>.
+     * Getter for <code>transactions.country</code>.
      */
     public String getCountry() {
         return (String) get(11);
     }
 
     /**
-     * Setter for <code>minance.transactions.zipcode</code>.
+     * Setter for <code>transactions.zipcode</code>.
      */
     public void setZipcode(String value) {
         set(12, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.zipcode</code>.
+     * Getter for <code>transactions.zipcode</code>.
      */
     public String getZipcode() {
         return (String) get(12);
     }
 
     /**
-     * Setter for <code>minance.transactions.amount</code>.
+     * Setter for <code>transactions.amount</code>.
      */
     public void setAmount(BigDecimal value) {
         set(13, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.amount</code>.
+     * Getter for <code>transactions.amount</code>.
      */
     public BigDecimal getAmount() {
         return (BigDecimal) get(13);
     }
 
     /**
-     * Setter for <code>minance.transactions.bank_name</code>.
+     * Setter for <code>transactions.bank_name</code>.
      */
     public void setBankName(String value) {
         set(14, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.bank_name</code>.
+     * Getter for <code>transactions.bank_name</code>.
      */
     public String getBankName() {
         return (String) get(14);
     }
 
     /**
-     * Setter for <code>minance.transactions.account_name</code>.
+     * Setter for <code>transactions.account_name</code>.
      */
     public void setAccountName(String value) {
         set(15, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.account_name</code>.
+     * Getter for <code>transactions.account_name</code>.
      */
     public String getAccountName() {
         return (String) get(15);
     }
 
     /**
-     * Setter for <code>minance.transactions.upload_time</code>.
+     * Setter for <code>transactions.upload_time</code>.
      */
     public void setUploadTime(String value) {
         set(16, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.upload_time</code>.
+     * Getter for <code>transactions.upload_time</code>.
      */
     public String getUploadTime() {
         return (String) get(16);
     }
 
     /**
-     * Setter for <code>minance.transactions.is_duplicate</code>.
+     * Setter for <code>transactions.is_duplicate</code>.
      */
-    public void setIsDuplicate(Boolean value) {
+    public void setIsDuplicate(Integer value) {
         set(17, value);
     }
 
     /**
-     * Getter for <code>minance.transactions.is_duplicate</code>.
+     * Getter for <code>transactions.is_duplicate</code>.
      */
-    public Boolean getIsDuplicate() {
-        return (Boolean) get(17);
+    public Integer getIsDuplicate() {
+        return (Integer) get(17);
     }
 
     // -------------------------------------------------------------------------
@@ -290,12 +290,12 @@ public class TransactionsRecord extends UpdatableRecordImpl<TransactionsRecord> 
     // -------------------------------------------------------------------------
 
     @Override
-    public Row18<Integer, Integer, String, String, String, LocalDate, LocalDate, String, String, String, String, String, String, BigDecimal, String, String, String, Boolean> fieldsRow() {
+    public Row18<Integer, Integer, String, String, String, LocalDate, LocalDate, String, String, String, String, String, String, BigDecimal, String, String, String, Integer> fieldsRow() {
         return (Row18) super.fieldsRow();
     }
 
     @Override
-    public Row18<Integer, Integer, String, String, String, LocalDate, LocalDate, String, String, String, String, String, String, BigDecimal, String, String, String, Boolean> valuesRow() {
+    public Row18<Integer, Integer, String, String, String, LocalDate, LocalDate, String, String, String, String, String, String, BigDecimal, String, String, String, Integer> valuesRow() {
         return (Row18) super.valuesRow();
     }
 
@@ -385,7 +385,7 @@ public class TransactionsRecord extends UpdatableRecordImpl<TransactionsRecord> 
     }
 
     @Override
-    public Field<Boolean> field18() {
+    public Field<Integer> field18() {
         return Transactions.TRANSACTIONS.IS_DUPLICATE;
     }
 
@@ -475,7 +475,7 @@ public class TransactionsRecord extends UpdatableRecordImpl<TransactionsRecord> 
     }
 
     @Override
-    public Boolean component18() {
+    public Integer component18() {
         return getIsDuplicate();
     }
 
@@ -565,7 +565,7 @@ public class TransactionsRecord extends UpdatableRecordImpl<TransactionsRecord> 
     }
 
     @Override
-    public Boolean value18() {
+    public Integer value18() {
         return getIsDuplicate();
     }
 
@@ -672,13 +672,13 @@ public class TransactionsRecord extends UpdatableRecordImpl<TransactionsRecord> 
     }
 
     @Override
-    public TransactionsRecord value18(Boolean value) {
+    public TransactionsRecord value18(Integer value) {
         setIsDuplicate(value);
         return this;
     }
 
     @Override
-    public TransactionsRecord values(Integer value1, Integer value2, String value3, String value4, String value5, LocalDate value6, LocalDate value7, String value8, String value9, String value10, String value11, String value12, String value13, BigDecimal value14, String value15, String value16, String value17, Boolean value18) {
+    public TransactionsRecord values(Integer value1, Integer value2, String value3, String value4, String value5, LocalDate value6, LocalDate value7, String value8, String value9, String value10, String value11, String value12, String value13, BigDecimal value14, String value15, String value16, String value17, Integer value18) {
         value1(value1);
         value2(value2);
         value3(value3);
@@ -714,7 +714,7 @@ public class TransactionsRecord extends UpdatableRecordImpl<TransactionsRecord> 
     /**
      * Create a detached, initialised TransactionsRecord
      */
-    public TransactionsRecord(Integer transactionId, Integer accountId, String category, String description, String transactionType, LocalDate transactionDate, LocalDate postDate, String memo, String address, String city, String stateName, String country, String zipcode, BigDecimal amount, String bankName, String accountName, String uploadTime, Boolean isDuplicate) {
+    public TransactionsRecord(Integer transactionId, Integer accountId, String category, String description, String transactionType, LocalDate transactionDate, LocalDate postDate, String memo, String address, String city, String stateName, String country, String zipcode, BigDecimal amount, String bankName, String accountName, String uploadTime, Integer isDuplicate) {
         super(Transactions.TRANSACTIONS);
 
         setTransactionId(transactionId);
